pyproject.toml
[project]
name = "onderwijsvl-python-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dyslexic-Goegrapher", email = "arnomoerdijk@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
